Align the file summary
Use <table> element to lay out the file summary. The summary is tabular
data anyway so the use of <table> should be justified. First column
contains lines added, justified to right and the second column contains
lines removed, justified to the left. The last column has the file name.

Fixes #27