Replace space to &nbsp; for HTML Rendering.
Reason:
Some javascript frameworks(ex. vue.js) manipulate HTML tags when drawing.

At that time, if there is a space-only DOM element, the space in the DOM
element is deleted.

Before:
`` `
<span> </span>
`` `

After (space deleted)
`` `
<span></span>
`` `

In order to avoid this, it has been modified to render with &nbsp; only when there is no diff.