Make d2h-d-none override visibility
Since `.d2h-d-none` is the first rule in the file, other rules with the same specificity override it, e.g. line 102 in `.d2h-files-diff`. This means hiding no longer works. This commit fixes that; alternatively, the `.d2h-d-none` rule should be moved to the end of the file if you don't like `!important`.